requirements.txt
torch
torchvision
torchaudio
flask
Flask-Cors
flask-mail
google-generativeai
diffusers
transformers
accelerate
requests
pillow
numpy
opencv-python
scikit-learn
matplotlib
torchcam